requirements.txt
accessible-pygments==0.0.5
alabaster==0.7.16
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
Authlib==1.3.2
babel==2.16.0
badurls==0.11
base==0.0.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
boxes==0.0.0
breathe==4.35.0
bs4==0.0.2
cachelib==0.13.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==2.0.12
click==8.1.8
colorama==0.4.6
colored==2.2.4
cryptography==44.0.0
decorator==5.1.1
defusedcsv==2.0.0
defusedxml==0.7.1
dirsearch==0.4.3.post1
distro==1.9.0
Django==5.1.4
docutils==0.19
executing==2.1.0
fastapi==0.115.6
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gitignore_parser==0.1.11
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-ntlm==1.4.0
idna==3.10
imagesize==1.4.1
ipython==8.31.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.5
jiter==0.8.2
keyring==25.6.0
linkify-it-py==2.0.3
lolcat==1.4
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
more-itertools==10.5.0
msgspec==0.19.0
mysql-connector-python==9.1.0
myst-parser==3.0.1
nh3==0.2.20
ntlm-auth==1.5.0
numpy==1.26.4
numpydoc==1.8.0
openai==1.57.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
phonenumbers==8.13.52
pkginfo==1.12.0
plotly==5.24.1
polib==1.2.0
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg-binary==3.2.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydata-sphinx-theme==0.16.1
Pygments==2.18.0
pyOpenSSL==24.3.0
pyparsing==3.2.0
PySocks==1.7.1
pyspnego==0.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
readme_renderer==44.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
requests_ntlm==1.3.0
rfc3986==2.0.0
rich==13.9.4
SecretStorage==3.3.3
setuptools==75.7.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==6.2.0
sphinx-autobuild==2024.10.3
sphinx-copybutton==0.5.2
sphinx-lint==1.0.0
sphinx_design==0.6.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-mermaid==1.0.0
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.41.3
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
tqdm==4.67.1
traitlets==5.14.3
twine==6.0.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
uvicorn==0.34.0
watchdog==3.0.0
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.1.3
wheel==0.45.1
wheelfile==0.0.9
writer==0.8.2
writer-sdk==1.5.0
WTForms==3.2.1
